fix: address senior developer feedback
✅ Architectural improvements:
- Pass full CreateAnalysisDto instead of extracting uploadId (analysis.controller.ts:98)
- Move BloodworkJobData interface to dedicated types file
- Remove artificial delays from analysis processor for real-world testing
- Extract helper functions to pure utilities in analysis/utils/

✅ Type safety enhancements:
- Replace 'any' with strongly-typed PatientInfo interface
- Create BloodworkParseResult response interface for better type safety
- Add comprehensive type definitions and barrel exports

✅ Naming convention fixes:
- Implement consistent DTO (input) / Response (output) suffix pattern
- Rename ApiResponseDto → ApiEnvelope to avoid NestJS decorator conflicts
- Update all imports and method signatures accordingly

✅ Security improvements:
- Add missing ParseFilePipeBuilder validation to uploads controller
- Implement defense-in-depth with explicit size validation
- Fix potential DoS vulnerability from missing endpoint validation

✅ Development workflow:
- Add dynamic IP detection using Expo Constants for network resilience
- Remove hardcoded IP addresses for automatic network adaptation

All feedback items addressed with working end-to-end flow and enhanced security.
