+++ bloodwork-backend/src/analysis/utils/index.ts
@@ -1,0 +1,21 @@
+/**
+ * Analysis Utils - Barrel Export
+ * 
+ * WHY: This barrel export file provides a clean way to import all
+ * analysis utility functions from a single location. It keeps imports
+ * organized and makes the utility API surface more discoverable.
+ * 
+ * USAGE:
+ * import { extractPatientName, parseTestResults, determineTestType } from '../utils';
+ * // Instead of individual imports from each file
+ */
+
+// Text extraction utilities
+export { extractPatientName, extractTestDate } from './text-extraction';
+
+// Test parsing utilities  
+export { parseTestResults, determineTestStatus } from './test-parsing';
+export type { TestStatus } from './test-parsing';
+
+// Test classification utilities
+export { determineTestType } from './test-classification';
